refactor: project menu 데이터 상수화
