+++ frontend/manage/src/components/pages/ProjectDetail/index.tsx
@@ -1,11 +1,11 @@
 import { FormEvent, useEffect } from "react";
 import { Redirect, useHistory, useRouteMatch } from "react-router-dom";
-import { ROUTE } from "../../../constants";
+import { PROJECT_MENU, ROUTE } from "../../../constants";
 import { useEditProject, useGetProject, useInput } from "../../../hooks";
 import { useDeleteProject } from "../../../hooks/useDeleteProject";
 import ScreenContainer from "../../../styles/ScreenContainer";
+import { MenuType } from "../../../types/menu";
 import { isEmptyString } from "../../../utils/validation";
-import ProjectSideBar from "../../organisms/ProjectSideBar";
 import SideBarTemplate from "../../organisms/SideBarTemplate";
 import {
   Container,
@@ -84,7 +84,7 @@ const ProjectDetail = () => {
 
   return (
     <ScreenContainer>
-      <SideBarTemplate projectId={projectId}>
+      <SideBarTemplate menus={PROJECT_MENU.get(projectId)}>
         <Container>
           <Title>프로젝트 정보</Title>
           <Form onSubmit={onEditProject}>
