+++ frontend/manage/src/constants/index.ts
@@ -4,3 +4,4 @@ export { REACT_QUERY_KEY } from "./reactQueryKey";
 export { BASE_URL, QUERY } from "./api";
 export { DOMAIN } from "./domain";
 export { GUIDE_FILE } from "./guideFile";
+export { PROJECT_MENU } from "./menu";
